git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@549 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ src/util/stringSupport/string_functions.C
@@ -1180,7 +1180,7 @@ StringUtility::escapeNewLineCharaters ( const string & X )
      return returnString;
    }
 
-std::string StringUtility::intToHex(uint32_t i) {
+std::string StringUtility::intToHex(uint64_t i) {
   ostringstream os;
   os << "0x" << std::hex << i;
   return os.str();
